website-search-backend/requirements.txt
absl-py==1.0.0
altair==4.2.2
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
autopep8==2.0.1
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.6.2
blis==0.7.11
branca==0.4.2
cachetools==5.3.0
caer==2.0.8
catalogue==2.0.10
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cloudpathlib==0.16.0
colorama==0.4.4
colorama-lpa @ file:///Users/anjalikedia/Downloads/colorama_lpa-0.4.4b1.0-py2.py3-none-any.whl
comm==0.2.2
confection==0.1.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.8.0
dlib==19.23.0
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fastjsonschema==2.19.1
filelock==3.13.4
Flask==2.1.1
Flask-Cors==4.0.0
flatbuffers==24.3.25
folium==0.12.1.post1
fonttools==4.39.2
fqdn==1.5.1
frozendict==2.4.2
fsspec==2024.3.1
future==0.18.3
gast==0.5.3
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.34.3
idna==3.3
imageio==2.16.1
importlib-metadata==6.6.0
imutils==0.5.4
ipykernel==6.29.4
ipython==8.21.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
keras==2.15.0
kiwisolver==1.4.4
lazr.uri==1.0.6
libclang==18.1.1
lxml==5.2.1
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
murmurhash==1.0.10
mypy==1.0.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.7.1
nltk==3.8.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.6.0
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
openpyxl==3.1.5
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.2
pandas==2.0.1
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.3
pexpect==4.9.0
Pillow==9.0.1
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
platformdirs==4.2.0
preshed==3.0.9
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pygame==2.1.2
pyglet==1.5.22
Pygments==2.15.1
PyJWT==2.3.0
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.1.1
python-json-logger==2.0.7
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
PyYAML==6.0.1
pyzmq==26.0.2
ratelim==0.1.6
referencing==0.34.0
regex==2024.4.16
requests==2.32.4
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.19.2
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.3
sentence-transformers==5.0.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.47.2
streamlit==1.22.0
sympy==1.13.1
tenacity==8.2.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorstore==0.1.45
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.17.1
tf-slim==1.1.0
tf_keras==2.15.1
thinc==8.2.3
threadpoolctl==3.4.0
tifffile==2024.2.12
tinycss2==1.2.1
tokenizers==0.21.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.3.3
tqdm==4.66.2
traitlets==5.7.1
traittypes==0.2.1
transformers==4.54.1
tweepy==4.14.0
twilio==7.7.0
typer==0.9.4
types-python-dateutil==2.9.0.20240316
types-pytz==2024.1.0.20240417
types-setuptools==69.5.0.20240415
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.35.0
validators==0.20.0
vega-datasets==0.9.0
virtualenv==20.25.3
wadllib==1.3.6
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==3.6.6
wordcloud==1.9.3
wrapt==1.14.1
xarray==2023.7.0
xarray-einstats==0.7.0
xgboost==2.0.3
xlrd==2.0.1
xyzservices==2024.4.0
yarl==1.9.4
yellowbrick==1.5
yfinance==0.2.38
zict==3.0.0
zipp==3.18.1
